Add code to fetch data and update view
This commit adds code that connects the app to the Firestore database,
fetches all the tasks of the user, and updates the table view in
TasksViewController.

A new struct is added to ensure that data is easily managed throughout
the app, and functions to convert the data to and from the Firebase JSON
structure and the struct are partially added (only convertJSONToTask)
for now.