Lay out foundation code for onboarding
This commit lays out the foundation for new user onboarding. This commit
adds a new function that fetches the configuration status of the user (a
manually-set Boolean value); if false, then the user will be redirected
to onboarding.

This commit also adds the Onboarding Storyboard; while empty now, future
commits will populate the folder with more files and screens.