📝 Update README
This commit makes some changes to the README file to improve the
readability; it makes the README more understandable to the average
user!

It seems like the README may be too cramped up as well. I may consider
moving some chunks of the README into a collated wiki section using
GitHub's wiki feature.