🚑 Fix Focus Mode bug
I'm sorry, but I really wanted the world to burn after this one. This
commit should close #6.

I realised that there was an issue with Focus Mode in the fact that if
the user tries to execute another one immediately after one has
concluded, the entire thing just... didn't work. I'm still not exactly
certain on what the issue is behind it, but after five hours of
scratching my head and staring intensely at the screen, I added one (1)
variable to FMTimerViewController and it just worked.

I'm not at the right mind to question this right now, but if it works,
it works?

If you're reading this and you're able to contribute to the project,
please do! I may need some help here on this bit while I figure this
out.