✨ Add support for localisation
This commit enables support for localisation in the project. Now, I'll
be able to have contributors in translation (that is, if anyone's even
interested, to begin with...)! I've added the Malay language as a
translation in here, but I'm not 100% confident in it.

Perhaps I'll set up some kind of translation platform soon, maybe?