🐛 Prevent device sleeping in Focus Mode
This commit makes some changes to the behaviour of
FMTimerViewController. Now, the view controller will prevent the phone
screen from sleeping - and subsequently, preventing the movement of the
app to the background. I'm not exactly sure what will happen once the
app is in the background, but I'm certain that timers wouldn't work as
they should. Therefore, I'm implementing this change!