🚧 Add basic account page
This commit adds the basic foundation design for AccountViewController.
Currently, it includes basic details of the user's account that are the
most important, such as the user's account ID, e-mail address, and the
number of tasks. There's also a button to delete all data (tasks) and
the account; a future to-do!